+++ Retail/Code/Retail/Retail_BE/Tables/AccountPackageRecurCharge.sql
@@ -1,6 +1,6 @@
 ﻿CREATE TABLE [Retail_BE].[AccountPackageRecurCharge] (
     [ID]                    BIGINT           IDENTITY (1, 1) NOT NULL,
-    [AccountPackageID]      INT              NOT NULL,
+    [AccountPackageID]      BIGINT           NOT NULL,
     [ChargeableEntityID]    UNIQUEIDENTIFIER NOT NULL,
     [BalanceAccountTypeID]  UNIQUEIDENTIFIER NULL,
     [BalanceAccountID]      VARCHAR (50)     NULL,
@@ -13,6 +13,8 @@
     [CreatedTime]           DATETIME         CONSTRAINT [DF_AccountPackageRecurCharge_CreatedTime] DEFAULT (getdate()) NULL,
     CONSTRAINT [PK_AccountPackageRecurCharge] PRIMARY KEY CLUSTERED ([ID] ASC)
 );
+
+
